requirements.txt
captum==0.7.0
clip==1.0
h5py==3.12.1
imageio==2.35.1
matplotlib==3.6.3
numpy==2.3.5
openai_clip==1.0.1
opencv_python==4.10.0.84
pandas==1.5.3
Pillow==12.0.0
scikit_learn==1.3.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
skimage==0.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.0
transformers==4.53.1
util==0.0.1
